Pronouns display support. Different colour than other fields.
Not sure if I want to have the fields separate entirely. Resizing the fonts is a pain. Currently pronouns are set to be the same font size no matter what, but this looks bad I think.
Maybe I should look into tech for  dynamically resizing pronouns alongside the other text?